Creating a BMP image from a UML CLass
We can now create a BMP image from a class signature (i.e. class name + list of member variables). Next we create the code necessary to parse through a file and create class dynamically.

+++ taswira/Geometry.h
@@ -44,7 +44,7 @@ namespace taswira {
             int DeltaCol = col2 - col1;
             int DeltaRow = row2 - row1;
             int Result = (DeltaCol * DeltaCol) + (DeltaRow * DeltaRow);
-            Result = std::sqrt(Result);
+            Result = static_cast<int>(std::sqrt(Result));
 
             return Result;
         }
